requirements.txt
openai>=1.0.0
google-generativeai>=0.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
schedule>=1.2.0
python-dotenv>=1.0.0
html2text>=2020.1.16
selenium>=4.15.0
webdriver-manager>=4.0.0



